requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiosqlite==0.19.0
pillow==10.1.0
pydantic==2.5.2
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
reportlab==4.0.7

